WLT-834: moved rootdomain to genesisrefs
