+++ insolar/genesisrefs/refs.go
@@ -53,6 +53,24 @@ var PredefinedPrototypes = map[string]insolar.Reference{
 	insolar.GenesisNameWallet + PrototypeSuffix:                *GenerateProtoReferenceFromContractID(PrototypeType, insolar.GenesisNameWallet, 0),
 }
 
+func init() {
+	for _, el := range insolar.GenesisNameMigrationDaemonMembers {
+		PredefinedPrototypes[el+PrototypeSuffix] = *GenerateProtoReferenceFromContractID(PrototypeType, insolar.GenesisNameMember, 0)
+	}
+
+	for _, el := range insolar.GenesisNameMigrationDaemons {
+		PredefinedPrototypes[el+PrototypeSuffix] = *GenerateProtoReferenceFromContractID(PrototypeType, insolar.GenesisNameMigrationDaemon, 0)
+	}
+
+	for _, el := range insolar.GenesisNamePublicKeyShards {
+		PredefinedPrototypes[el+PrototypeSuffix] = *GenerateProtoReferenceFromContractID(PrototypeType, insolar.GenesisNamePKShard, 0)
+	}
+
+	for _, el := range insolar.GenesisNameMigrationAddressShards {
+		PredefinedPrototypes[el+PrototypeSuffix] = *GenerateProtoReferenceFromContractID(PrototypeType, insolar.GenesisNameMigrationShard, 0)
+	}
+}
+
 var (
 	// ContractRootDomain is the root domain contract reference.
 	ContractRootDomain = GenesisRef(insolar.GenesisNameRootDomain)
